+++ lint-janitor.config.js
@@ -1,7 +1,6 @@
 module.exports = {
   esFiles: ['test/**/*.js', 'test/**/*.ts', 'src/**/*.ts'],
-  sassFiles: ['test/**/*.scss'],
-  styleFiles: ['test/**/*.less'],
+  styleFiles: ['test/**/*.less', 'test/**/*.scss'],
   htmlFiles: ['test/**/*.html'],
   jsonFiles: ['*.json', '.eslintrc', '.htmlhintrc', 'lint/**/*.json', 'lint/.eslintrc', 'lint/.htmlhintrc', 'test/*.json', 'src/**/*.json'],
   yamlFiles: ['./.travis.yml', './.sass-lint.yml', './lint/.sass-lint.yml']