+++ src/cli.ts
@@ -10,7 +10,6 @@ import { ZConfigNullReader } from './lint-config/config-null-reader.class';
 import { ZConfigReader } from './lint-config/config-reader.class';
 import { IZLintArgs } from './lint/lint-args.interface';
 import { ZLint } from './lint/lint.class';
-import { ZSassLint } from './sass-lint/sass-lint.class';
 import { ZStyleLint } from './style-lint/style-lint.class';
 import { ZYamlLint } from './yaml-lint/yaml-lint.class';
 
@@ -26,9 +25,9 @@ const yamlLint = new ZYamlLint();
 
 const zLint = new ZLint(logger);
 zLint.esLint = new ZEsLint(logger);
-zLint.sassLint = new ZSassLint(logger);
 zLint.styleLint = new ZStyleLint(logger);
 zLint.htmlHint = new ZFileLint(htmlHint, jsonConfigReader, logger, 'html');
 zLint.jsonLint = new ZFileLint(jsonLint, nullConfigReader, logger, 'json');
 zLint.yamlLint = new ZFileLint(yamlLint, nullConfigReader, logger, 'yaml');
+
 zLint.run(args).then((result) => (process.exitCode = result));