+++ src/lint/lint-options.interface.ts
@@ -1,11 +1,9 @@
 export interface IZLintOptions {
   esConfig?: string;
-  sassConfig?: string;
   styleConfig?: string;
   htmlConfig?: string;
 
   esFiles?: string[];
-  sassFiles?: string[];
   styleFiles?: string[];
   htmlFiles?: string[];
   jsonFiles?: string[];